Big commit: VistA UTF-8 support + improvements
List of manual tests I have performed:
OTF = OSEHRA Testing Framework; pi = post-install)
- OSEHRA VistA (no arguments)
- OSEHRA VistA (OTF + UTF-8 + YDB + Panorama + pi)
- OSEHRA VistA (Skip OTF + UTF-8 + YDB)
- OSEHRA VistA (PIP SQL Mapping)
- WorldVistA   (Skip OTF + YDB + Panorama)
- vxVistA      (Skip OTF + YDB + pi)
- VEHU         (Skip OTF + GTM + Panorama)
- RPMS         (Skip OTF + YDB + pi)
- Cache        (pi)
- Cache        (extract)

LIST OF CHANGES
===============

README
------
- Add Plan VI version
- Update WV version with new release
- Remove references to `./Common/removeVistASource.sh`. This is now done
  automatically by the installer.

Dockerfile
----------
- Add recode to installed programs list. Use recode rather than iconv since
  recode can convert files in place. Conversion was necessary for routines,
  some of which contained ISO-8859-1 characters.
- Remove old inaccurate comments from Docker file.

autoInstaller.sh
----------------
- -u support to install VistA with UTF-8
- -z support to turn on set -x and leave downloaded files behind rather than
  deleting them.
- Get VistA-Dashboard code from OSEHRA-Sandbox as old code in OSEHRA repo force
  C locale.
- All data intensive operations will utilize multiple cores if available--these
  are: routine compilation, routine iconv, sed on globals, chmod.
- Virgin install Kernel routines updated to vXU-8.0-10002.
- Importer should not start Taskman. There is no point in doing so as we will
  stop it one second later.
- Routine compilation used to be done twice - here for OSEHRA Testing Framework,
  and in GTM/importVistA.sh for import. Del from the latter and just do it here
  for both pathways.

GTM/createVistAInstance.sh
--------------------------
- -u flag to install VistA with UTF-8 on YDB/GTM
- Remedy some missing environment variables
-- gtm_lct_stdnull to force standard null collation for local variables.
-- gtm_sysid should be the instance name. Has only a cosemtic effect for
   various users.
-- gtm_zstep to make stepping nicer.
-- gtm_link to link routines even if they are on the stack--the changes will
   apply when the stack exits that routine.
-- gtm_side_effects & gtm_boolean are NOT set intentionally--I want to sunset
   their use in other scripts I have if we really find out they are not needed.
-- set gtm_etrap appropriately for prog.sh and tied.sh. This is the inital error
   trap until VistA sets one.
- Update the list of non-journaled/temp globals per VistA documentation
- Comments/80C cleanup

GTM/install.sh
--------------
Fix wrong comment.

GTM/importVistA.sh
------------------
- Support for $devMode flag
- Routine copying is now done with find/xargs for speed
- Remove routine compilation here and move to autoInstaller.sh. See comment
  there.
- Comments/80C cleanup

Common/ov6pi.sh & Common/ov6piko.m
----------------------------------
Post installer scripts for Plan VI docker image

+++ autoInstaller.sh
@@ -70,9 +70,11 @@ usage()
       -q    Install SQL mapping for YottaDB
       -s    Skip testing
       -w    Install RPMS XINETD scripts
+      -u    Install GTM/YottaDB with UTF-8 enabled
       -y    Use YottaDB
       -v    Build ViViaN Documentation
       -x    Extract given M[UMPS] code
+      -z    Dev Mode: Don't clean-up and set -x
 
     NOTE:
     The Caché install only supports using .DAT files for the VistA DB, and
@@ -82,7 +84,7 @@ usage()
 EOF
 }
 
-while getopts ":ha:cbxemdgi:vp:sr:wyq" option
+while getopts ":ha:cbxemdugi:vp:sr:wyqz" option
 do
     case $option in
         h)
@@ -126,6 +128,9 @@ do
         s)
             skipTests=true
             ;;
+        u)
+            utf8=true
+            ;;
         v)
             generateViVDox=true
             ;;
@@ -142,6 +147,9 @@ do
         q)
             developmentDirectories=true
             installSQL=true
+            ;;
+        z)
+            devMode=true
             ;;
     esac
 done
@@ -211,6 +219,18 @@ if [ -z $installSQL ]; then
     installSQL=false;
 fi
 
+if [ -z $utf8 ]; then
+    utf8=false;
+fi
+
+if [ -z $devMode ]; then
+    devMode=false
+fi
+
+if $devMode; then
+    set -x
+fi
+
 # Quit if no M environment viable
 if [[ ! $installgtm || ! $installcache || ! $installYottaDB ]]; then
     echo "You need to either install Caché, GT.M or YottaDB!"
@@ -230,6 +250,7 @@ echo "Skip bootstrap: $bootstrap"
 echo "Use Cache: $installcache"
 echo "Use GT.M: $installgtm"
 echo "Use YottaDB: $installYottaDB"
+echo "GT.M/YDB in UTF-8: $utf8"
 echo "Install RPMS scripts: $installRPMS"
 echo "Running on local repo: $localVistARepo"
 
@@ -265,7 +286,7 @@ fi
 if ! $skipTests; then
     cd /usr/local/src
     rm -rf VistA-Dashboard
-    git clone -q https://github.com/OSEHRA/VistA -b dashboard VistA-Dashboard
+    git clone -q https://github.com/OSEHRA-Sandbox/VistA -b dashboard VistA-Dashboard
 fi
 
 # See if vagrant folder exists if it does use it. if it doesn't clone the repo
@@ -318,6 +339,9 @@ fi
 if $installRPMS; then
    createVistaInstanceOptions+="-r "
 fi
+if $utf8; then
+   createVistaInstanceOptions+="-u "
+fi
 
 if $installgtm || $installYottaDB ; then
     cd GTM
@@ -353,6 +377,11 @@ else
     USER_HOME=/root
 fi
 
+# number of cores - 1
+# (works also on MacOS; on FreeBSD, omit underscore)
+cores=$(($(getconf _NPROCESSORS_ONLN) - 1))
+if (( cores < 1 )); then cores=1; fi
+
 # source env script during running user's login
 if $installgtm || $installYottaDB; then
     echo "source $basedir/etc/env" >> $USER_HOME/.bashrc
@@ -361,7 +390,6 @@ fi
 if (($installgtm || $installYottaDB) && ! $generateViVDox); then
 
   echo "Getting the VistA-M Source Code"
-  # Clone VistA-M repo
   pushd /usr/local/src
   if [[ $repoPath == *.git ]]; then
       if ! [ -z $branch ]; then
@@ -377,6 +405,16 @@ if (($installgtm || $installYottaDB) && ! $generateViVDox); then
       rm VistA-M-master.zip
       mv $dir VistA-Source
   fi
+
+  # Make routines/globals importable if UTF-8
+  if $utf8; then
+      echo "Modifying zwr globals to contain UTF-8 in the first line"
+      find VistA-Source -name '*.zwr' -print0 | xargs -0 -I{} -n 1 -P $cores \
+        sed -i '1c\OSEHRA ZGO Export: THIS GLOBAL UTF-8' "{}"
+      echo "Convert non-ASCII routines to UTF-8"
+      find VistA-Source -name '*.m' -print0 | xargs -0 -I{} -n 1 -P $cores \
+        recode -f iso8859-1..utf-8 "{}"
+  fi
   popd
 
   if $skipTests; then
@@ -384,10 +422,12 @@ if (($installgtm || $installYottaDB) && ! $generateViVDox); then
       cd $basedir
 
       # Perform the import
+      export devMode       # Send this guy down
       su $instance -c "source $basedir/etc/env && $scriptdir/GTM/importVistA.sh"
+      export -n devMode    # and not any further!
 
       # Get GT.M Optimized Routines from Kernel-GTM project and unzip
-      curl -fsSLO --progress-bar https://github.com/shabiel/Kernel-GTM/releases/download/XU-8.0-10001/virgin_install.zip
+      curl -fsSLO --progress-bar https://github.com/shabiel/Kernel-GTM/releases/download/XU-8.0-10002/virgin_install.zip
 
       # Unzip file, put routines, delete old objects
       su $instance -c "unzip -qo virgin_install.zip -d $basedir/r/"
@@ -400,9 +440,6 @@ if (($installgtm || $installYottaDB) && ! $generateViVDox); then
 
       # Run the auto-configurer accepting the defaults
       su $instance -c "source $basedir/etc/env && mumps -run START^KBANTCLN"
-
-      # Start Taskman
-      su $instance -c "source $basedir/etc/env && cd ~/tmp/ && mumps -run ^ZTMB"
   else
       # Build a dashboard and run the tests to verify installation
       # These use the Dashboard branch of the VistA repository
@@ -424,21 +461,18 @@ if (($installgtm || $installYottaDB) && ! $generateViVDox); then
 
       # Import VistA and run tests using OSEHRA automated testing framework
       su $instance -c "source $basedir/etc/env && ctest -S $scriptdir/test.cmake -V"
+
       # Tell users of their build id
       echo "Your build id is: $buildid you will need this to identify your build on the VistA dashboard"
-
-      # Compile routines
-      # number of cores - 1
-      # (works also on MacOS; on FreeBSD, omit underscore)
-      cores=$(($(getconf _NPROCESSORS_ONLN) - 1))
-      if (( cores < 1 )); then cores=1; fi
-
-      echo "Compiling routines"
-      cd $basedir/r/$gtmver
-      find .. -name '*.m' | xargs --max-procs=$cores --max-args=1 $gtm_dist/mumps >> $basedir/log/compile.log 2>&1
-      echo "Done compiling routines"
   fi
+
+  echo "Compiling routines"
+  cd $basedir/r/$gtmver
+  find .. -name '*.m' | xargs --max-procs=$cores --max-args=1 $gtm_dist/mumps >> $basedir/log/compile.log 2>&1
+  echo "Done compiling routines"
+
 fi
+
 # Enable journaling
 if $installgtm || $installYottaDB; then
     su $instance -c "source $basedir/etc/env && $basedir/bin/enableJournal.sh"
@@ -495,7 +529,7 @@ fi
 # Ensure group permissions are correct
 if $installgtm || $installYottaDB; then
     echo "Please wait while I fix the group permissions on the files..."
-    chmod -R g+rw /home/$instance
+    find /home/$instance -print0 | xargs -0 -I{} -n 1 -P $cores chmod g+rw "{}"
 fi
 
 extract=""
@@ -509,8 +543,11 @@ if $generateViVDox; then
 fi
 
 # Clean up the VistA-M source directories to save space
-if $skipTests; then
-    rm -rf /usr/local/src/VistA-Source
-else
-    rm -rf $basedir/Dashboard/VistA-M
+echo "Cleaning up..."
+if ! $devMode; then
+    if $skipTests; then
+        rm -rf /usr/local/src/VistA-Source
+    else
+        rm -rf $basedir/Dashboard/VistA-M
+    fi
 fi
